requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
Faker==19.12.0
fastapi==0.104.0
GeoAlchemy2==0.14.2
geographiclib==2.0
geopy==2.4.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
pydantic==2.4.2
pydantic-core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.23.2
zipp==3.17.0